pyproject.toml
[tool.poetry]
name = "xmlparser"
version = "0.1.0"
description = ""
authors = ["defaultftxuser <banginlord@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.5"}
celery = "^5.4.0"
asyncpg = "^0.30.0"
alembic = "^1.14.0"
motor = "^3.6.0"
pytest-asyncio = "^0.24.0"
faker = "^33.0.0"
lxml = "^5.3.0"
redis = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
